+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - New locales
   ([#585](https://github.com/chriso/validator.js/pull/585))
+- Added support for greater or less than in `isFloat()`
+  ([#544](https://github.com/chriso/validator.js/issues/544))
 
 #### 6.0.0